requirements.txt
appdirs==1.4.4
asgiref==3.3.4
Babel==2.9.1
blis==0.7.4
catalogue==2.0.6
certifi==2020.12.5
cffi==1.14.6
chardet==3.0.4
click==7.1.2
DAWG-Python==0.7.2
distlib==0.3.3
docopt==0.6.2
extract-msg==0.23.1
Flask==1.1.1
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
future==0.18.2
googlefinance==0.7
gunicorn==20.1.0
idna==2.10
IMAPClient==2.1.0
intervaltree==3.1.0
ipymarkup==0.9.0
isodate==0.6.0
itsdangerous==2.0.1
joblib==1.0.1
kiwisolver==1.3.2
Mako==1.1.4
msrest==0.6.21
msrestazure==0.6.4
murmurhash==1.0.5
natasha==1.4.0
navec==0.10.0
ndg-httpsclient==0.5.1
NRCLex==3.0.0
numpy==1.19.4
oauthlib==3.1.1
packaging==20.9
pandas==1.1.4
pathspec==0.9.0
pathy==0.4.0
pbr==1.6.0
pdfminer.six==20181108
pefile==2021.9.3
preshed==3.0.5
prettytable==2.1.0
psycopg2==2.8.6
py-moneyed==2.0
pyasn1==0.4.8
pybind11==2.6.2
pycorenlp==0.3.0
pycparser==2.20
pycryptodome==3.9.9
pydantic==1.7.3
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
pywin32-ctypes==0.2.0
razdel==0.5.0
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.6
scikit-learn==0.23.2
scipy==1.5.4
six==1.16.0
slovnet==0.5.0
smart-open==3.0.0
sortedcontainers==2.4.0
sqlparse==0.4.1
srsly==2.4.1
textblob==0.15.3
thinc==8.0.3
threadpoolctl==2.2.0
tqdm==4.62.3
typer==0.3.2
typing-extensions==3.10.0.2
tzlocal==1.5.1
urllib3==1.25.11
virtualenv==20.4.6
vk==2.0.2
waitress==1.4.4
wasabi==0.8.2
torch==1.9.0
torchvision==0.11.1
Pillow==8.0.1
efficientnet-pytorch==0.7.1